+++ sgl-router/src/routers/grpc/common/mod.rs
@@ -1,0 +1,5 @@
+//! Shared code for both regular and harmony routers
+
+pub mod response_collection;
+pub mod response_formatting;
+pub mod stages;
