+++ sgl-router/src/routers/grpc/common/response_formatting.rs
