+++ sgl-router/src/routers/grpc/common/stages/dispatch_metadata.rs
@@ -5,7 +5,7 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use async_trait::async_trait;
 use axum::response::Response;
 
-use crate::routers::grpc::stages::PipelineStage;
+use super::PipelineStage;
 use crate::routers::grpc::{
     context::{DispatchMetadata, RequestContext, RequestType, WorkerSelection},
     error,
