+++ sgl-router/src/routers/grpc/common/stages/worker_selection.rs
@@ -6,7 +6,7 @@ use async_trait::async_trait;
 use axum::response::Response;
 use tracing::warn;
 
-use crate::routers::grpc::stages::PipelineStage;
+use super::PipelineStage;
 use crate::{
     core::{ConnectionMode, Worker, WorkerRegistry, WorkerType},
     policies::PolicyRegistry,
