+++ sgl-router/src/routers/grpc/harmony/stages/preparation.rs
@@ -12,9 +12,9 @@ use crate::{
         responses::ResponsesRequest,
     },
     routers::grpc::{
+        common::stages::PipelineStage,
         context::{PreparationOutput, RequestContext, RequestType},
         error,
-        stages::PipelineStage,
         utils,
     },
 };
