+++ sgl-router/src/routers/grpc/harmony/stages/response_processing.rs
@@ -7,9 +7,10 @@ use axum::response::Response;
 
 use super::super::{HarmonyResponseProcessor, HarmonyStreamingProcessor};
 use crate::routers::grpc::{
+    common::stages::PipelineStage,
     context::{FinalResponse, RequestContext, RequestType},
+    utils,
     error,
-    stages::PipelineStage,
 };
 
 /// Harmony Response Processing stage: Parse and format Harmony responses
