+++ sgl-router/src/routers/grpc/regular/mod.rs
@@ -1,0 +1,8 @@
+//! Regular (non-harmony) model processing
+//!
+//! This module contains all code specific to regular tokenizer-based models,
+//! including pipeline stages, response processing, and streaming.
+
+pub mod processing;
+pub mod stages;
+pub mod streaming;
