+++ sgl-router/src/routers/grpc/regular/stages/chat/response_processing.rs
@@ -1,26 +1,29 @@
-//! Generate response processing stage: Handles both streaming and non-streaming responses
+//! Chat response processing stage: Handles both streaming and non-streaming responses
+//!
+//! - For streaming: Spawns background task and returns SSE response (early exit)
+//! - For non-streaming: Collects all responses and builds final ChatCompletionResponse
 
-use std::{sync::Arc, time::Instant};
+use std::sync::Arc;
 
 use async_trait::async_trait;
 use axum::response::Response;
 
 use crate::routers::grpc::{
+    common::stages::PipelineStage,
     context::{FinalResponse, RequestContext},
-    processing, streaming,
-    stages::PipelineStage,
+    regular::{processing, streaming},
     utils,
 };
 
-/// Generate response processing stage
+/// Chat response processing stage
 ///
-/// Extracts generate-specific response processing logic from the old unified ResponseProcessingStage.
-pub struct GenerateResponseProcessingStage {
+/// Extracts chat-specific response processing logic from the old unified ResponseProcessingStage.
+pub struct ChatResponseProcessingStage {
     processor: processing::ResponseProcessor,
     streaming_processor: Arc<streaming::StreamingProcessor>,
 }
 
-impl GenerateResponseProcessingStage {
+impl ChatResponseProcessingStage {
     pub fn new(
         processor: processing::ResponseProcessor,
         streaming_processor: Arc<streaming::StreamingProcessor>,
@@ -33,22 +36,21 @@ impl GenerateResponseProcessingStage {
 }
 
 #[async_trait]
-impl PipelineStage for GenerateResponseProcessingStage {
+impl PipelineStage for ChatResponseProcessingStage {
     async fn execute(&self, ctx: &mut RequestContext) -> Result<Option<Response>, Response> {
-        self.process_generate_response(ctx).await
+        self.process_chat_response(ctx).await
     }
 
     fn name(&self) -> &'static str {
-        "GenerateResponseProcessing"
+        "ChatResponseProcessing"
     }
 }
 
-impl GenerateResponseProcessingStage {
-    async fn process_generate_response(
+impl ChatResponseProcessingStage {
+    async fn process_chat_response(
         &self,
         ctx: &mut RequestContext,
     ) -> Result<Option<Response>, Response> {
-        let start_time = Instant::now();
         let is_streaming = ctx.is_streaming();
 
         // Extract execution result
@@ -70,17 +72,18 @@ impl GenerateResponseProcessingStage {
         if is_streaming {
             // Streaming: Use StreamingProcessor and return SSE response (done)
             return Ok(Some(
-                self.streaming_processor.clone().process_streaming_generate(
+                self.streaming_processor.clone().process_streaming_response(
                     execution_result,
-                    ctx.generate_request_arc(), // Cheap Arc clone (8 bytes)
+                    ctx.chat_request_arc(), // Cheap Arc clone (8 bytes)
                     dispatch,
                 ),
             ));
         }
 
         // Non-streaming: Delegate to ResponseProcessor
-        let request_logprobs = ctx.generate_request().return_logprob.unwrap_or(false);
-        let generate_request = ctx.generate_request_arc();
+        let request_logprobs = ctx.chat_request().logprobs;
+
+        let chat_request = ctx.chat_request_arc();
 
         let stop_decoder = ctx
             .state
@@ -89,20 +92,19 @@ impl GenerateResponseProcessingStage {
             .as_mut()
             .ok_or_else(|| utils::internal_error_static("Stop decoder not initialized"))?;
 
-        let result_array = self
+        let response = self
             .processor
-            .process_non_streaming_generate_response(
+            .process_non_streaming_chat_response(
                 execution_result,
-                generate_request,
+                chat_request,
                 dispatch,
                 stop_decoder,
                 request_logprobs,
-                start_time,
             )
             .await?;
 
         // Store the final response
-        ctx.state.response.final_response = Some(FinalResponse::Generate(result_array));
+        ctx.state.response.final_response = Some(FinalResponse::Chat(response));
 
         Ok(None)
     }
