+++ sgl-router/src/routers/grpc/regular/stages/generate/mod.rs
